added Count() method to base Repo Interface; working on update, delete, insert functions

+++ Peek.WebService/Models/Result.cs
@@ -10,6 +10,6 @@ namespace Peek.WebService.Models
     public class Result<T>
     {
         public List<T> Records { get; set; }
-        public int TotalRecordsCount { get; set; }
+        public long TotalRecordsCount { get; set; }
     }
 }
